Partly fix race with inotify watching newly created directories.
"mkdir -p a/b/c/d" results in a CREATE for "a", but mkdir makes the
rest before inotifywatch can start watching "a". We can fix these
simple (and probably common) cases of the race by adding everything
in "a" to the list of files when we see "a" get created.